requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
crispy_bootstrap4
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms
django_extensions
django-heroku==0.3.1
djangorestframework
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.2
mypy-extensions==0.4.3
openpyxl
pandas
pathspec==0.9.0
psycopg2==2.9.1
pytz==2021.1
regex==2021.8.28
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
tomli==1.2.1
urllib3==1.26.6
whitenoise==5.3.0
django-import-export
